requirements.txt
pdfplumber==0.10.3
openai==1.3.7
streamlit==1.28.1
reportlab==4.0.7
pytest==7.4.3
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
nltk==3.8.1 